% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{draw}
\alias{draw}
\title{Draws Fireflow Objects}
\usage{
draw(fireflow, color = "black", LineType = 1)
}
\arguments{
\item{fireflow}{fireflow object defined using ff function.}

\item{color}{A specification for the default plotting color. See section ‘Color Specification’.}

\item{LineType}{The line type. Line types can either be specified as an integer (0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash, 5=longdash, 6=twodash)}
}
\description{
Draws Fireflow Objects
}
\examples{
n185()

a <- ff(60,3000,20, "first")
b <- ff(50,2750, 35, "second")

draw(a, "red", 1)
draw(b, "blue", 2)

draw_legend()
}
\seealso{
\code{\link{ff}}
\code{\link{tilt}}
\code{\link{shift}}
}
