% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{draw_legend}
\alias{draw_legend}
\title{Draw Legend}
\usage{
draw_legend(position = "topright")
}
\arguments{
\item{position}{keyword for position.  See Details.

The location may be specified by setting position to a single keyword from the list "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". This places the legend on the inside of the plot frame at the given location. Partial argument matching is used.}
}
\description{
Draw Legend
}
\examples{
n185()

a <- ff(60,3000,20, "first")
b <- ff(50,2750, 35, "second")

draw(a, "red", 1)
draw(b, "blue", 2)

draw_legend()
}
