% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ff}
\alias{ff}
\title{Creates a fire flow supply curve based on a single hydrant test per NFPA 291 Recommended Practice for Fire Flow Testing}
\usage{
ff(Ps, Qt, Pt, ID = "")
}
\arguments{
\item{Ps}{Static Pressure from Field Fire Flow Test, in PSI}

\item{Qt}{Test Flow from Field Fire Flow Test, in GPM}

\item{Pt}{Test Residual from Field Fire Flow Test, in GPM}

\item{ID}{Unique name for Field Fire Flow Test (used for legend)}
}
\value{
FireFlow Object
}
\description{
Creates a fire flow supply curve based on a single hydrant test per NFPA 291 Recommended Practice for Fire Flow Testing
}
\examples{
#' n185()

a <- ff(60,3000,20, "first")
b <- ff(50,2750, 35, "second")

draw(a, "red", 1)
draw(b, "blue", 2)

draw_legend()
}
