% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{kp}
\alias{kp}
\title{Unit Friction Slope}
\usage{
kp(D, C = 130)
}
\arguments{
\item{D}{Pipe diameter in Inches}

\item{C}{Hazen-Williams friction factor}
}
\value{
k' (numerical) unit friction slope for 1 LF of pipeline (corrected for PSI)
}
\description{
Unit Friction Slope
}
\examples{
# friction factor for 5,000 LF of 12-inch pipe
k <- 5000*kp(12)

# Headloss at 2,000 GPM for 5,000 LF of 12-inch pipe
k*2000^1.85

#' # friction factor for 5,000 LF of 12-inch pipe at C=100
k <- 5000*kp(12, 100)

# Headloss at 2,000 GPM for 5,000 LF of 12-inch pipe
k*2000^1.85
}
