% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{n185}
\alias{n185}
\title{Provides a framework for N^1.85 logarithmic graph based on NFPA 291 recommendations.}
\usage{
n185(MaxFlow = 5000)
}
\arguments{
\item{MaxFlow}{Maximum Flow in GPM}
}
\value{
framework for drawing fireflow objects
}
\description{
Provides a framework for N^1.85 logarithmic graph based on NFPA 291 recommendations.
}
\examples{
n185()

a <- ff(60,3000,20)

draw(a)
}
