% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pt}
\alias{pt}
\title{Draw Fire-Flow Point}
\usage{
pt(Qt = NULL, Pt = NULL, fireflow = NULL, color = "black")
}
\arguments{
\item{Qt}{Test Flow to be plotted, in GPM}

\item{Pt}{Test Residual to be plotted, in GPM}

\item{fireflow}{fireflow object defined using ff function.}

\item{color}{A specification for the point plotting color. See section ‘Color Specification’.}
}
\description{
Test Test Test
}
\examples{
n185()

a <- ff(60,3000,20, "first")
b <- ff(50,2750, 35, "second")

draw(a, "red", 1)
draw(b, "blue", 2)

draw_legend()

pt(3000,,a, "red")
pt(,20,"blue")
pt(2000,20)
}
