% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shift}
\alias{shift}
\title{Title}
\usage{
shift(fireflow, static)
}
\arguments{
\item{static}{}
}
\description{
Title
}
\examples{
n185()

a <- ff(55, 4500, 40)
draw(a)

b <- shift(a, 40)
draw(b, "grey50", 2)
}
