% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tilt}
\alias{tilt}
\title{Tilt FireFlow Friction Slope}
\usage{
tilt(fireflow, friction_slope)
}
\arguments{
\item{fireflow}{fireflow object, defined using ff()}

\item{friction_slope}{proposed reduction in friction slope}
}
\value{
FireFlow Object
}
\description{
Tilt FireFlow Friction Slope
}
\examples{
n185()

a <- ff(55, 4500, 40)
draw(a)

# add friction loss from 1000 LF of 12-inch pipeline
b <- tilt(a, 1000*kp(12))
draw(b, "grey50", 2)
}
