% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{n185}
\alias{n185}
\title{Create a new n185 plot.}
\usage{
n185(MaxFlow = 5000, MaxPressure = 100)
}
\arguments{
\item{MaxFlow}{Maximum Flow in GPM for x-axis}

\item{MaxPressure}{Maximum Pressure in PSI for y-axis}
}
\value{
framework for drawing fireflow objects
}
\description{
Provides a framework for \out{n<sup>1.85</sup>} logarithmic graph based on NFPA 291 recommendations.
}
\examples{
n185()

a <- ff(60,3000,20)

draw(a)
}
